requirements.txt
Django==4.0.4
djangorestframework==3.13.1
pip==21.1.1
pytz==2022.1
sqlparse==0.4.2
mysqlclient==2.1.0
asgiref==3.5.0
setuptools==56.0.0